requirements.txt
streamlit
pandas
selenium
python-dotenv
beautifulsoup4
mysql-connector-python
